+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 16 December 2024 - v1.0.1 ##
+* create landing page template
+* script header footer
+
 ## 11 December 2024 - v1.0.0 ##
 
 * Initial Files